Working with latest cores library
This has several improvements. The most critical to make it work with the teensy cores library was the addition of the file in the build list.
Separately, this also moves the setting of compiler flags to the toolchain file, so that all teensy files being built will have access.
It also includes support for uploading to teensy via the CLI (windows application called from WSL Ubuntu).